+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 ### Added
 - support for RMarkdown (.Rmd) in script directives.
 
+### Changed
+- Benchmark files now also include the maximal RSS and VMS size of the Snakemake process and all sub processes.
+
+
 ## [3.11.2] - 2017-03-15
 ### Changed
 - Fixed fix handling of local URIs with the wrapper directive.
-
 
 
 ## [3.11.1] - 2017-03-14